update work_rel test

+++ back-work/src/main/java/kr/co/dbinc/back_work/model/Work_RelVO.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 public class Work_RelVO {
 
 	private Integer workID;
-	private Integer workerID;	
+	private Long workerID;	
 }
